New code style (#21)
* Update zyper.c

Use my new code style.

* Update zyper.h

Use size_t for indexing variables.

* Create Makefile

Add a Makefile to make compiling easier.

* Update README.md

Use make to use the Makefile.

+++ zyper.h
@@ -4,7 +4,7 @@
 struct _zyper_data
 {
 	char **wrds;
-	short wrds_amount;
+	size_t wrds_amount;
 };
 
 typedef struct _zyper_data zyper_data;
